style:언어 이미지 svg 수정
